move wifi into own lib
